restore int64() assertion

+++ go/logic/migrator.go
@@ -1291,7 +1291,7 @@ func (this *Migrator) executeWriteFuncs() error {
 						if niceRatio := this.migrationContext.GetNiceRatio(); niceRatio > 0 {
 							copyRowsDuration := time.Since(copyRowsStartTime)
 							sleepTimeNanosecondFloat64 := niceRatio * float64(copyRowsDuration.Nanoseconds())
-							sleepTime := time.Duration(sleepTimeNanosecondFloat64) * time.Nanosecond
+							sleepTime := time.Duration(int64(sleepTimeNanosecondFloat64)) * time.Nanosecond
 							time.Sleep(sleepTime)
 						}
 					}
